+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `pycmAverageError` class
 ### Changed
 - Document modified
+- `relabel` method bug fixed
+- Minor bug in `Compare` class fixed
 ## [2.6] - 2020-03-25
 ### Added
 - `custom_rounder` function